Oops, we need to actually call the syscall for it to work. :)

+++ e3.1/test.pl
@@ -6,6 +6,11 @@ unless (-e $mod) {
     die "File <$mod> could not be found in the current directory.";
 } 
 
+my $res = system "kldload $mod";
+if ($res != 0) {
+    die "Kernel module failed to load correctly (exit $res).\n";
+} 
+
 system 'killall cat';
 
 my $ps = `ps`;
@@ -20,10 +25,7 @@ if ($ps !~ /cat/) {
     die "Could not start cat.\n";
 }
 
-my $res = system "kldload $mod";
-if ($res != 0) {
-    die "Kernel module failed to load correctly (exit $res).\n";
-} 
+syscall(210, "cat");
 
 my $ps = `ps`;
 if ($ps =~ /cat/) {
